==> 00_preprosessing_sql.Rmd <==
---
title: 'Green Regional Path paper: Preprocessing (SQL)'
author: "Daniel S. Hain"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_notebook:
    df_print: paged
    toc: yes
    toc_depth: 3
    toc_float: yes
    number_sections: yes
    code_folding: hide
---

## Setup

```{r, setup, include=FALSE}
# Not sure if necessary... was an issue with a version conflict... try new version in the future again...
# devtools::install_version("RPostgres", version = "1.2.1", repos = "http://cran.us.r-project.org")

### general options
Sys.setenv(LANG = "en")
options("scipen" = 100, "digits" = 4) # override R's tendency to use scientific notation

### Clean workspace
rm(list=ls())
graphics.off()

### Load packages (maybe need to be installed first)
# Standard
library(tidyverse) # General DS toolkit
library(magrittr) # For advanced piping

# Databases
library(DBI) # GEneral R database interface
library(RPostgres) # PostgreSQL interface driver 
library(dbplyr) # for dplyr with databases
```

### Connect to database

```{r}
# Load script with variables
source("../variables.R")

# set up connection to existing PostgreSQL database, just plug in own details
con <- dbConnect(drv = RPostgres::Postgres(), 
                 dbname = var_dbname,
                 host = var_host, 
                 port = var_port,
                 user = var_user, 
                 password = var_password,
                 sslmode = 'require'
                 )

rm(var_dbname, var_host, var_password, var_port, var_user)
```

```{r}
con %>% dbListTables() %>% sort()
```

# Defining parameters

```{r}
var_t_start <- 1985
var_t_end <- 2015
var_t_break <- 2000
```

# Main Data

## Loading tables

```{r}
tbl_appln <- tbl(con, 'tls201_appln')
tbl_appln_prior <- tbl(con, 'tls204_appln_prior')
tbl_pers_appln <- tbl(con, 'tls207_pers_appln') 
tbl_person <- tbl(con, 'tls206_person')
tbl_appln_cpc <- tbl(con, 'tls224_appln_cpc')
tbl_docdb_fam_cpc <- tbl(con, 'tls225_docdb_fam_cpc')
tbl_isic_cpc <- tbl(con, 'xtra_isic_cpc')

```

# Regions

```{r}
select_region_nordic <- tibble(
  person_ctry_code = c('SE', 'NO','DK', 'FI'
  )) %>%
  copy_to(con, ., "xtra_select_region", 
          temporary = TRUE, overwrite = TRUE)
#tbl_select_region <- tbl(con, 'xtra_select_region')

# # or select by nuts
# nuts = c(
#     'SE232', # Västra Götaland county
#     'NO043', # Rogaland 
#     'DK012' # Copenhagen suburbs))
```

```{r}
# select persons in regions
tbl_region_person <- tbl_person %>%
  filter(nuts_level == 3) %>%
  #semi_join(tbl_select_region, by = 'person_ctry_code') %>%
  compute()
```

```{r}
# Identify person-applications
tbl_region_pers_appln <- tbl_pers_appln %>%
  filter(invt_seq_nr > 0) %>%
  group_by(appln_id) %>%
  mutate(n_inv = n()) %>%
  ungroup() %>%
  mutate(weight_frac = 1/n_inv) %>%
  inner_join(tbl_region_person %>% select(person_id, person_ctry_code, nuts), by = 'person_id') %>%
  compute()
```



```{r}
# Identify applications
tbl_region_appln <- tbl_appln %>%
  # filter
  filter(granted == 'Y',
         appln_filing_year >= var_t_start,
         appln_filing_year <= var_t_end)  %>%
  # Restrict to EU
  semi_join(tbl_region_pers_appln, by = 'appln_id')  %>%
  # Reduce to PAtent family
  group_by(docdb_family_id) %>%
    slice_min(order_by = appln_filing_date, n = 1, with_ties = FALSE) %>%
  ungroup() %>%
  # Insert period
  compute()
```

```{r}
# Update person dataframe for only inventors 
tbl_region_person  %<>%
  semi_join(tbl_region_pers_appln, by = 'person_id') %>%
  compute()
```


# Technologies

```{r}
# Filter CPCs in region and label Y02 tags
tbl_region_docdb_fam_cpc <- tbl_docdb_fam_cpc %>%
  semi_join(tbl_region_appln, by = 'docdb_family_id') %>%
  distinct(docdb_family_id, cpc_class_symbol) %>%
  mutate(Y_tag = (cpc_class_symbol %like% 'Y02%') | (cpc_class_symbol %like% 'Y04S%')) %>%
  compute()
```

```{r}
# Identify person-applications
tbl_region_appln %<>%
  left_join(tbl_region_docdb_fam_cpc %>% 
              filter(Y_tag == TRUE) %>% 
              distinct(docdb_family_id, Y_tag), 
            by = 'docdb_family_id') %>%
  replace_na(list(Y_tag = FALSE))  %<>%
  compute()
```

```{r}
tbl_region_appln %>% count(Y_tag)
```
```{r}
# Update person dataframe for only inventors (with Y-tag)
tbl_region_pers_appln   %<>%
  inner_join(tbl_region_appln %>% select(appln_id, Y_tag), by = 'appln_id') %>%
  compute()
```


# Industries

```{r}
tbl_appln_isic <- tbl_region_docdb_fam_cpc %>%
  select(docdb_family_id, cpc_class_symbol) %>%
  semi_join(tbl_region_appln, by = 'docdb_family_id') %>%
  mutate(cpc3 = cpc_class_symbol %>% str_sub(1, 3)) %>%
  inner_join(tbl_isic_cpc, by = 'cpc3') %>%
  # Select one level higher
  mutate(isic4_l3 = isic4_l4 %>% str_sub(1, 3)) %>%
  group_by(docdb_family_id, isic4_l3) %>%
  summarise(probability_weight = sum(probability_weight, na.rm = TRUE)) %>%
  ungroup() %>%
  compute()
```

# Applicants

Note: Not applicants from the region, but applicants that have inventors in the region
Note: Applicants patents to region via inventor

```{r}
# Now get the applicants of all the applcations

# Identify person-applications
tbl_region_applt_appln <- tbl_pers_appln %>%
  filter(applt_seq_nr > 0) %>% # Only applicant
  semi_join(tbl_person %>% # Only nordics
              semi_join(select_region_nordic, by = 'person_ctry_code'),
            by = 'person_id') %>%
  inner_join(tbl_region_appln %>% select(appln_id, Y_tag), by = 'appln_id') %>%
  group_by(appln_id) %>%
  mutate(n_applt = n()) %>%
  ungroup() %>%
  mutate(weight_frac = 1/n_applt) %>%
  compute()
```

```{r}
# Identify applicants
tbl_region_applt <- tbl_person  %>%
  semi_join(tbl_region_applt_appln, by = 'person_id') %>%
  compute()
```



# Save all

```{r}
tbl_region_appln %>% collect() %>% write_rds('../temp/tbl_region_appln.rds')

tbl_region_pers_appln %>% collect() %>% write_rds('../temp/tbl_region_pers_appln.rds')
tbl_region_person %>% collect() %>% write_rds('../temp/tbl_region_person.rds')

tbl_region_applt_appln %>% collect() %>% write_rds('../temp/tbl_region_applt_appln.rds')
tbl_region_applt %>% collect() %>% write_rds('../temp/tbl_region_applt.rds')

tbl_region_docdb_fam_cpc %>% collect() %>% write_rds('../temp/tbl_region_docdb_fam_cpc.rds')
tbl_appln_isic %>% collect() %>% write_rds('../temp/tbl_appln_isic.rds')
```

# close



```{r}
dbDisconnect(con)
```

```{r}
sessionInfo()
```
