==> 01_calc_specialization.Rmd <==
---
title: 'Green Regional Path paper: Constructing Technology Spaces'
author: "Daniel S. Hain"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_notebook:
    df_print: paged
    toc: yes
    toc_depth: 3
    toc_float: yes
    number_sections: yes
    code_folding: hide
---

## Setup

```{r, setup, include=FALSE}
### general options
Sys.setenv(LANG = "en")
options("scipen" = 100, "digits" = 4) # override R's tendency to use scientific notation

### Clean workspace
rm(list=ls())
graphics.off()

### Load packages (maybe need to be installed first)
# Standard
library(tidyverse) # General DS toolkit
library(magrittr) # For advanced piping

# Databases
library(DBI) # GEneral R database interface
library(RPostgres) # PostgreSQL interface driver 
library(dbplyr) # for dplyr with databases

# Load functions
source("functions/functions_relatedness.R")

### Connect to database
# Load script with variables
source("../variables.R")

# set up connection to existing PostgreSQL database, just plug in own details
con <- dbConnect(drv = RPostgres::Postgres(),  dbname = var_dbname, host = var_host,  port = var_port, user = var_user,  password = var_password, sslmode = 'require')

rm(var_dbname, var_host, var_password, var_port, var_user)
```

```{r}
# con %>% dbListTables() %>% sort()
```

```{r}
tbl_nuts <- tbl(con, 'tls904_nuts')
```

# Defining parameters

```{r}
var_t_start <- 1985
var_t_end <- 2015
var_t_break <- 2000
```

# Load tables

```{r}
data_appln <- read_rds('../temp/tbl_region_appln.rds')
data_docdb_fam_cpc <- read_rds('../temp/tbl_region_docdb_fam_cpc.rds')
data_appln_isic <- read_rds('../temp/tbl_appln_isic.rds')
data_pers_appln <- read_rds('../temp/tbl_region_pers_appln.rds')
#data_person <- read_rds('../temp/tbl_region_person.rds') 
```

# Initial data cleaning

```{r}
data_appln %<>% mutate(period = ((appln_filing_year >= var_t_break) + 1) %>% as.numeric() %>% as.character()) 
```


# Technology space (By technology classes)

```{r}
list_isic  <- read_csv('../input/list_isic4_cleaned.csv') 
```

## Create Relatedness matrix

```{r}
# devtools::install_github("PABalland/EconGeo", force = T)
library(EconGeo)
```

```{r}
mat_tech <- create_sparse_matrix(i = data_appln_isic %>% pull(docdb_family_id),
                                       j = data_appln_isic %>% pull(isic4_l3),
                                       weight = data_appln_isic %>% pull(probability_weight),
                                       projection = 'j',
                                       sparse = FALSE)
```

```{r}
mat_tech %<>% relatedness(method = "cosine")
```

## Technology Space

```{r}
library(tidygraph)
library(ggraph)
```

```{r}
g_tech <- mat_tech %>% as_tbl_graph(directed = FALSE) %N>%
  left_join(list_isic %>% filter(level == 3) %>% select(isic4_l3, isic4_l3_name, isic4_l2_name, isic4_l1_name), by = c('name' = 'isic4_l3')) %>%
  mutate(dgr = centrality_eigen(weights = weight)) %E>%
  filter(weight >= mean(weight))
```

```{r}
g_tech %N>% as_tibble() 
```

## Regional specialization

```{r}
region_tech <- data_appln %>%
  select(appln_id, docdb_family_id, appln_filing_year, period, Y_tag) %>%
  left_join(data_appln_isic %>% select(docdb_family_id, isic4_l3, probability_weight), by = 'docdb_family_id') %>%
  left_join(data_pers_appln %>% select(appln_id, person_id, nuts, n_inv), by = 'appln_id') %>%
  mutate(weight_frac = probability_weight / n_inv)
```

# Regional path identification

```{r}
region_RTA <- region_tech %>%  
  drop_na(Y_tag, nuts, isic4_l3, period) %>%
  # Summing weights of non-Y and Y tagged 
  group_by(Y_tag, nuts, isic4_l3, period) %>%
  summarise(n_tech_region = sum(weight_frac, na.rm = TRUE)) %>%
  ungroup() %>%
  # Calculate location quotient
    group_by(Y_tag, period) %>%
    mutate(n = sum(n_tech_region, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(Y_tag, period, nuts) %>%
    mutate(n_region = sum(n_tech_region, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(Y_tag, period, isic4_l3) %>%
    mutate(n_tech = sum(n_tech_region, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(rta = ((n_tech_region / n_region) / (n_tech / n)) %>% round(2),
         rta_bin = (rta >= 1) %>% as.numeric() ) %>%
  # Replace NAs
  replace_na(list(rta = 0, rta_bin = 0, n_region = 0, n_tech = 0, n_tech_region = 0, n = 0)) %>%
  complete(Y_tag, nuts, isic4_l3, period, fill = list(rta = 0, rta_bin = 0, n_region = 0, n_tech = 0, n_tech_region = 0, n = 0))
```

# Save all

```{r}
# Adittional ones
g_tech %>% write_rds('../temp/g_tech.rds')
region_RTA %>% write_rds('../temp/region_RTA.rds')
region_tech %>% write_rds('../temp/region_tech.rds')
```

```{r}
sessionInfo()
```
