==> 99_adddon_rr3.Rmd <==
---
title: 'Updated stuff for RR3'
author: "Daniel S. Hain"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_notebook:
    df_print: paged
    toc: yes
    toc_depth: 3
    toc_float: yes
    number_sections: yes
    code_folding: hide
---

## Setup

```{r, setup, include=FALSE}
# Not sure if necessary... was an issue with a version conflict... try new version in the future again...
# devtools::install_version("RPostgres", version = "1.2.1", repos = "http://cran.us.r-project.org")

### general options
Sys.setenv(LANG = "en")
options("scipen" = 100, "digits" = 4) # override R's tendency to use scientific notation

### Clean workspace
rm(list=ls())
graphics.off()

### Load packages (maybe need to be installed first)
# Standard
library(tidyverse) # General DS toolkit
library(magrittr) # For advanced piping

# Databases
library(DBI) # GEneral R database interface
library(RPostgres) # PostgreSQL interface driver 
library(dbplyr) # for dplyr with databases
```

# ISIC classification table

```{r}
data <- read_csv2('../input/list_isic4.csv')
```

```{r}
data %<>%
  mutate(code = ifelse(letter %in% c('A', 'B') & !is.na(as.numeric(code)), paste0('0', code), code),
         level = str_length(code)) %>%
  rename(isic4_l1 = letter) %>%
  mutate(isic4_l2 = ifelse(level == 2, code, NA),
         isic4_l3 = ifelse(level == 3, code, NA),
         isic4_l4 = ifelse(level == 4, code, NA)) %>%
  group_by(isic4_l1) %>% fill(isic4_l2, .direction = 'down') %>% ungroup() %>% 
  group_by(isic4_l2) %>% fill(isic4_l3, .direction = 'down') %>% ungroup() %>% 
  select(isic4_l1, isic4_l2, isic4_l3, isic4_l4, level, description) %>%
  rename(isic4_l4_name = description) %>%
  mutate(isic4_l3_name = ifelse(level == 3, isic4_l4_name, NA)) %>%
  mutate(isic4_l2_name = ifelse(level == 2, isic4_l4_name, NA)) %>%
  mutate(isic4_l1_name = ifelse(level == 1, isic4_l4_name, NA)) %>%
  group_by(isic4_l1) %>% fill(isic4_l1_name, .direction = 'down') %>% ungroup() %>%
  group_by(isic4_l2) %>% fill(isic4_l2_name, .direction = 'down') %>% ungroup() %>%
  group_by(isic4_l3) %>% fill(isic4_l3_name, .direction = 'down') %>% ungroup() 
```

```{r}
data
```


```{r}
data %>% pull(isic4_l2) %>% n_distinct()
```



```{r}
data %>% write_csv('../input/list_isic4_cleaned.csv')
rm(data)
```

# ISIC4 CPC concordance
```{r}
# Load script with variables
source("../variables.R")

# set up connection to existing PostgreSQL database, just plug in own details
con <- dbConnect(drv = RPostgres::Postgres(), 
                 dbname = var_dbname,
                 host = var_host, 
                 port = var_port,
                 user = var_user, 
                 password = var_password,
                 sslmode = 'require'
                 )

rm(var_dbname, var_host, var_password, var_port, var_user)
```

```{r}
data <- read_csv('../input/list_isic4_cpc.csv') %>%
  rename(isic4_l4 = isic_rev4_4)
```

```{r}
data %>% copy_to(dest = con, .,
                 temporary = FALSE,
                 name = 'xtra_isic_cpc',
                 indexes = list('isic4_l4', 'cpc3'))
```

```{r}
tbl(con, 'xtra_isic_cpc')
```
